+++ testHW1.py
@@ -1,9 +1,8 @@
 import pytest
-from HW_1_OOP import Triangle
-from HW_1_OOP import Circle
+from HW1 import Triangle, Circle, TriangleException, CircleException
 from typing import Union
 
-side_transfer_test = (3, 5, 8), (5.0, 4.0, 8.0), (2.5, 9, 2)
+side_transfer_test = (3, 5, 8), (5.0, 4.0, 8.0), (6.5, 8, 4)
 
 
 @pytest.mark.parametrize('a, b, c', side_transfer_test)
@@ -16,20 +15,20 @@ def test_side_transfer(a: Union[float, int], b: Union[float, int], c: Union[floa
 
 @pytest.mark.xfail(raises=Exception)
 def test_error_zero_triangle():
-    with pytest.raises(Exception):
+    with pytest.raises(TriangleException):
         Triangle(0, 4, 6.9)
 
-    with pytest.raises(Exception):
+    with pytest.raises(TriangleException):
         Triangle(5, -5, 6.9)
 
 
 @pytest.mark.xfail(raises=Exception)
 def test_error_value_triangle():
 
-    with pytest.raises(Exception):
+    with pytest.raises(TriangleException):
         Triangle(5, "Koshmar", 6.9)
 
-    with pytest.raises(Exception):
+    with pytest.raises(TriangleException):
         Triangle(5, 5, "-6.9")
 
 
@@ -41,7 +40,7 @@ def test_triangle_square(triangle: Triangle, expect: float):
     assert triangle.square() == expect
 
 
-triangle_perimetr = [(Triangle(5, 5, 5), 15), (Triangle(9, 2, 5), 16), (Triangle(12, 10, 10), 32)]
+triangle_perimetr = [(Triangle(5, 5.0, 5), 15.0), (Triangle(9, 4.5, 5), 18.5), (Triangle(12, 10, 10), 32)]
 
 
 @pytest.mark.parametrize('triangle, expect', triangle_perimetr)
@@ -49,24 +48,15 @@ def test_triangle_perimetr(triangle: Triangle, expect: float):
     assert triangle.perimetr() == expect
 
 
-circle_test = [5.0, 2, 8.0]
-
-
-@pytest.mark.parametrize('radius', circle_test)
-def test_circle(radius: float):
-    circle = Circle(radius)
-    assert circle.radius == radius
-
-
 @pytest.mark.xfail(raises=Exception)
 def test_err_negative_num_circle():
-    with pytest.raises(Exception):
+    with pytest.raises(CircleException):
         Circle(-1)
 
 
 @pytest.mark.xfail(raises=Exception)
 def test_err_data_type_circle():
-    with pytest.raises(Exception):
+    with pytest.raises(CircleException):
         Circle("ulitka")
